[spark] Change Main Class from App to Driver
