+++ spark/src/test/java/com/revature/DriverTest.java
@@ -1,0 +1,20 @@
+package com.revature;
+
+import static org.junit.Assert.assertTrue;
+
+import org.junit.Test;
+
+/**
+ * Unit test for simple Driver.
+ */
+public class DriverTest
+{
+    /**
+     * Rigorous Test :-)
+     */
+    @Test
+    public void shouldAnswerWithTrue()
+    {
+        assertTrue( true );
+    }
+}
